Move Differentiate to new file algebra_dif. Refactor step_up
